test(antigravity): clear the last three Windows failures in the provider suite
The `windows-unit` job still reported three failures under
tests/unit/providers/antigravity. None of them were defects in the code
under test; all three were fixtures asserting something other than their
subject.

Two are the same race. `runPrint` resolves the turn from inside its `try`
and unlinks the agy log in the `finally` that follows, so cleanup is still
in flight when the awaited promise settles, and the number of hops between
the two is not fixed - the `fs.promises` call lands on the threadpool. Both
`preserves the agy log file when the run fails and removes it on success`
and `resolves with an already-delivered answer when the CLI hangs after it`
waited exactly one `setImmediate` before asserting the file was gone.
Instrumenting the wait showed the unlink always happens, after 1 to 4 ticks
depending on how many suites share the worker: running the file alone
failed while running the whole directory passed. Replace the fixed barrier
with `waitForRemovedFile`, which polls the condition.

The third was a launch-shape assumption of the kind #74 removed elsewhere.
`recovers models from Antigravity settings when agy models stdout is empty`
read `--log-file` out of `mockedSpawn.mock.calls[0][1]` directly, so on
Windows - where the arguments are nested inside the `cmd.exe /d /s /c
"<line>"` wrapper - the flag was not found, the fixture skipped writing its
log and settings files, and `recoverAntigravityModelsFromSettings` fell
through to `getDefaultAntigravityAppDataDir`. That reads the real
`~/.gemini/antigravity-cli/settings.json` of whoever runs the suite: on a
machine that actually uses Antigravity the assertion then compared the
developer's configured model against the fixture's, and reported a model
ordering diff for a process-launch problem. Route the fixture through
`toAgyArgs` and assert the flag is present before the recovery runs, so
this can never silently become a test of the host's own configuration.

`toAgyArgs` and `parseWindowsShellCommandLine` now live in
tests/helpers/antigravityLaunchShape.ts rather than being copied into a
second suite; `launch-shape normalization helpers` keeps pinning them
against the real producer from its new location.

On Windows tests/unit/providers/antigravity goes from 3 failures to 0
(147 passing), and the full unit suite from 39 failures to 36 with no new
ones. Verified on Node 22.22.2 / npm 12.0.2.
